requirements.txt
fastapi
uvicorn
psycopg
python-dotenv
python-jose
passlib[bcrypt]
bcrypt